+++ packages/tools/utils/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - add `isEqualTo`
 
+### Any -> (Object -> Boolean)
+
+- add `hasValue`
+
 ### Array -> (Array -> Array)
 
 - add `pluckKeys`